add dialog options for better usage

+++ update.sh
@@ -1,9 +1,27 @@
+#!/bin/bash
+
+if [ $1 == "--help" ]; then
+    echo "Updates the prospero repo with the current host config"
+    echo " "
+    echo "update [argument]"
+    echo " "
+    echo "options:"
+    echo "xfce                      specific for xfce ~/.config/xfce files"
+    exit 0
+fi
+
+
 echo "Updating repo files"
 
-echo "rofi"
-cat ~/.config/rofi/config.rasi > ./setup/rofi-setup/config.rasi
+./update/aliases.sh
 
-echo "xfce4 config"
-cp -r ~/.config/xfce4/xfconf ./setup
-cp -r ~/.config/xfce4/terminal ./setup
+if [ -f ~/.config/rofi/config.rasi ]; then
+    echo "backup rofi config"
+    cat ~/.config/rofi/config.rasi > ./setup/rofi-setup/config.rasi
+fi
 
+if [ 1 == "xfce" ]; then
+    echo "backup xfce config"
+    cp -r ~/.config/xfce4/xfconf ./setup
+    cp -r ~/.config/xfce4/terminal ./setup
+fi
